requirements.txt
Flask
mss
PyAutoGUI
Pillow